chore: refactor user capacity table
