+++ src/mlpa/run.py
@@ -70,7 +70,7 @@ async def lifespan(app: FastAPI):
         app_attest_connected = True
 
         await litellm_pg.create_budget()
-        await litellm_pg.ensure_capacity_state()
+        await app_attest_pg.ensure_capacity_state()
 
         yield
     finally:
